Added selecting flow
