+++ salon_creator/ios/Runner/Info.plist
@@ -25,7 +25,7 @@
 			<string>Editor</string>
 			<key>CFBundleURLSchemes</key>
 			<array>
-				<string>com.googleusercontent.apps.911023633551-ikue0k2o9bpmf6cane2rn011g6dj2fiv</string>
+				<string>com.googleusercontent.apps.649411725532-110sj4lkvb4b7p4u9938b8pop9t1dmm0</string>
 			</array>
 		</dict>
 	</array>
